+++ src/pocketmine/entity/behavior/navigator/algorithms/PythagorasAlgorithm.php
@@ -26,7 +26,10 @@ use pocketmine\utils\navigator\Tile;
 
 class PythagorasAlgorithm implements DistanceAlgorithm{
 	
-	public function calculate(Tile $from, Tile $to) : int{
-		return sqrt(pow($to->x - $from->x, 2) + pow($to->y - $from->y, 2));
+	public function calculate(Tile $from, Tile $to) : float{
+		return sqrt(
+		    pow($to->x - $from->x, 2) +
+            pow($to->y - $from->y, 2)
+        );
 	}
 }
